refactor: replace the func file name
