+++ src/commands/ping.rs
