+++ src/commands/stream.rs
