+++ tests/test_config_command.rs
@@ -1,6 +1,6 @@
 use rus_key::command_factory::Command;
 use rus_key::db::{DataType, Db};
-use rus_key::func::config::ConfigCommand;
+use rus_key::commands::config::ConfigCommand;
 use rus_key::init;
 
 #[test]
