+++ tests/test_string_command.rs
@@ -1,9 +1,9 @@
 use rus_key::command_factory::Command;
 use rus_key::db::DataType;
 use rus_key::db::Db;
-use rus_key::func::expired::ExpiredCommand;
-use rus_key::func::string::StringCommand;
-use rus_key::func::utils::UtilsCommand;
+use rus_key::commands::expired::ExpiredCommand;
+use rus_key::commands::string::StringCommand;
+use rus_key::commands::utils::UtilsCommand;
 use std::error::Error;
 
 fn get_current_time() -> i64 {
