+++ tests/test_utils_command.rs
@@ -1,8 +1,8 @@
 use rus_key::command_factory::Command;
 use rus_key::db::Db;
-use rus_key::func::hashmap::HashMapCommand;
-use rus_key::func::string::StringCommand;
-use rus_key::func::utils::UtilsCommand;
+use rus_key::commands::hashmap::HashMapCommand;
+use rus_key::commands::string::StringCommand;
+use rus_key::commands::utils::UtilsCommand;
 
 fn set_key(db: &mut Db) {
     // set key
